requirements.txt
Sanic
redis
hiredis
cassandra-driver